+++ client/src/pages/user/_registerModal.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
-import { handleKakaoLogin } from '../../util/snsLoginLogic';
+import { handleKakaoLogin } from '../../util/authorization/handleSnsLogin';
 import alertPreparingService from '../../util/alertPreparingService';
 import { ModalBackdrop, ModalContainer } from '../../components/UI/S_Modal';
 import { S_Label } from '../../components/UI/S_Text';
